requirements.txt
selenium
beautifulsoup4
pillow
